+++ bioscout/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "2.0.0b4"
+__version__ = "2.0.0b6"
 
 from typing import TYPE_CHECKING
 if TYPE_CHECKING:  # editor autocomplete only — no runtime cost